+++ lib/vagrant-templated/command/init.rb
@@ -1,106 +1,110 @@
-require 'optparse'
-
 module Vagrant
   module Templated
     module Command
       class Init < Vagrant.plugin('2', :command)
 
         def execute
-          root_path = File.expand_path '../../../', File.dirname(File.expand_path(__FILE__))
-          templates_attributes = Dir.glob(File.expand_path("config/templates/attributes/*.yml", root_path)).collect do |template|
-            YAML.load_file(template)
-          end.reduce Hash.new, :merge
-
-          options = {}
-          opts = OptionParser.new do |o|
-            o.banner = 'Usage: vagrant templated init [options] <template>'
-
-            o.separator ''
-            o.separator 'Templates availables:'
-            o.separator ''
-            templates_attributes.keys.each do |template|
-              o.separator "     #{template}"
-            end
-
-            o.separator ''
-            o.separator 'Options:'
-            o.separator ''
-
-            o.on('-s', '--suffix SUFFIX', String,
-                 "Output suffix for Vagrantfile and Berksfile. '-' for stdout") do |suffix|
-              options[:suffix] = suffix
-            end
-
-            o.on('-f', '--force', 'Overwrite an existing box if it exists') do |f|
-              options[:force] = f
-            end
+          template, version, options = setup
+          return if template.nil?
+          @env.ui.info "vagrant-templated detected #{template} template", new_line: true
+          if version.nil? || version.empty?
+            version = Catalog.patch template, version
+            @env.ui.info "vagrant-templated didn't received a version, using largest available #{version}"
+          else
+            @env.ui.info "vagrant-templated detected #{version} version"
           end
 
-          argv = parse_options(opts)
-          return if !argv
-          if argv.empty? || argv.length > 2
-            raise Vagrant::Errors::CLIInvalidUsage,
-              help: opts.help.chomp
-          end
+          vagrantfile = Catalog.vagrantfile_for template, version
+          berksfile = Catalog.berksfile_for template, version
 
-          template = argv[0]
-          if argv.length == 2
-            options[:name] = argv[0]
-            template = argv[1]
-          end
+          if options[:cat]
+            @env.ui.output vagrantfile
+            @env.ui.output berksfile
+          else
+            files_root = @env.cwd
+            if options[:output]
+              @env.ui.info "vagrant-templated received output path #{options[:output]}"
+              output_path = Pathname.new options[:output]
+              files_root =  if output_path.relative?
+                output_path.expand_path files_root
+              else
+                output_path
+              end
+              @env.ui.info "vagrant-templated will write the files in: #{files_root}"
+            end
 
-          @env.ui.info("Detected template: #{template}", prefix: false, color: :yellow)
-          raise Vagrant::Errors::VagrantTemplatedOptionNotFound unless templates_attributes.keys.include? template
-          template_attributes = templates_attributes[template]
-
-          vagrantfile_save_path = nil
-          berksfile_save_path = nil
-          if options[:suffix] != "-"
-            vagrantfile_save_path = Pathname.new(['Vagrantfile', options[:suffix]].compact.join('.')).expand_path(@env.cwd)
+            vagrantfile_save_path = Pathname.new('Vagrantfile').expand_path files_root
             vagrantfile_save_path.delete if vagrantfile_save_path.exist? && options[:force]
-            raise Vagrant::Errors::VagrantfileTemplatedExistsError if vagrantfile_save_path.exist?
-            berksfile_save_path = Pathname.new(['Berksfile', options[:suffix]].compact.join('.')).expand_path(@env.cwd)
+            raise Vagrant::Templated::Errors::VagrantfileExistsError if vagrantfile_save_path.exist?
+
+            berksfile_save_path = Pathname.new('Berksfile').expand_path files_root
             berksfile_save_path.delete if berksfile_save_path.exist? && options[:force]
-            raise Vagrant::Errors::BerksfileTemplatedExistsError if berksfile_save_path.exist?
-          end
+            raise Vagrant::Templated::Errors::BerksfileExistsError if berksfile_save_path.exist?
 
-          vagrantfile = ERB.new File.read(File.expand_path('config/templates/files/Vagrantfile.erb', root_path)), nil, '-'
-          berksfile = ERB.new File.read(File.expand_path('config/templates/files/Berksfile.erb', root_path)), nil, '-'
-
-          contents = vagrantfile.result binding
-          if vagrantfile_save_path
             begin
               vagrantfile_save_path.open('w+') do |f|
-                f.write(contents)
+                f.write vagrantfile
               end
+              @env.ui.info "vagrant-templated created #{vagrantfile_save_path}"
             rescue Errno::EACCES
-              raise Vagrant::Errors::VagrantfileWriteError
+              raise Vagrant::Templated::Errors::VagrantfileWriteError
             end
 
-            @env.ui.info('Templated Vagranfile created successfully', prefix: false)
-          else
-            @env.ui.info('VAGRANTFILE', prefix: false, color: :green)
-            @env.ui.info(contents, prefix: false)
-          end
-
-          contents = berksfile.result binding
-          if berksfile_save_path
             begin
               berksfile_save_path.open('w+') do |f|
-                f.write(contents)
+                f.write berksfile
               end
+              @env.ui.info "vagrant-templated created #{berksfile_save_path}"
             rescue Errno::EACCES
-              raise Vagrant::Errors::BerksfileWriteError
+              raise Vagrant::Templated::Errors::BerksfileWriteError
             end
-
-            @env.ui.info('Templated Berksfile created successfully', prefix: false)
-          else
-            @env.ui.info('BERKSFILE', prefix: false, color: :green)
-            @env.ui.info(contents, prefix: false)
           end
-            @env.ui.info('Vagrant Templated finished successfully.', prefix: false, color: :green)
+
+          @env.ui.success 'vagrant-templated finished successfully'
           0
         end
+
+        private
+
+          def setup
+            options = {}
+            opts = OptionParser.new do |o|
+              o.banner = 'Usage: vagrant templated init [options] <template> [version]'
+
+              o.separator ''
+              o.separator 'Templates and versions availables:'
+              o.separator ''
+              Catalog.templates.each do |template|
+                o.separator "     #{template}: #{Catalog.versions_for(template).join(', ')}"
+              end
+              o.separator ''
+              o.separator 'If no version is provided, larger version will be used.'
+
+              o.separator ''
+              o.separator 'Options:'
+              o.separator ''
+
+              o.on('-o', '--output OUTPUT', String, 'Output path for files. Default to `.`') do |o|
+                options[:output] = o
+              end
+
+              o.on('-c', '--cat', 'Output Vagrantfile and Berksfile to stdout') do |c|
+                options[:cat] = c
+              end
+
+              o.on('-f', '--force', 'Overwrite existing Vagrantfile and Berksfile if they exists') do |f|
+                options[:force] = f
+              end
+            end
+            argv = parse_options opts
+            return if !argv
+            if !argv || argv.empty? || argv.length > 2
+              raise Vagrant::Errors::CLIInvalidUsage,
+                help: opts.help.chomp
+            end
+            template, version = argv
+            [template, version, options]
+          end
       end
     end
   end